Arpan's Commit - 14-.03.18

+++ app/src/main/java/com/perfect11/contest/dto/LiveLeaderboardDto.java
@@ -3,7 +3,8 @@ package com.perfect11.contest.dto;
 import java.io.Serializable;
 
 public class LiveLeaderboardDto implements Serializable {
-    public int points;
+
+    public float points;
     public int team_id;
     public String reference_id;
 }
